feat(io): fix overlay, scene_emit, caf_transfer, scene_to_step f32/f64 boundaries
- overlay.rs: Coordinate3Node PVec3→Vec3 conversion
- scene_emit.rs: AnnotationPoint::local f64→f32 cast
- caf_transfer.rs: [u32;16]→[u64;16] hash key
- scene_to_step.rs: PVec3 push conversion

16 errors remaining: cross-crate boundary casts in rc3d-io

+++ crates/rc3d-io/src/step/caf_transfer.rs
@@ -286,7 +286,7 @@ fn build_xde_labels(
         tree.flatten_shells()
     };
 
-    let mut seen_placements: HashSet<(SolidKey, [u32; 16])> = HashSet::new();
+    let mut seen_placements: HashSet<(SolidKey, [u64; 16])> = HashSet::new();
 
     for (shell_step_id, world) in placements {
         let Some(sk) = dedup.resolve_solid(shell_step_id, shell_to_solid)? else {
